Add a Data Type Archive for Gimex
This archive contains structure information for the Graphics I/O
library used by SimCity 4.
The types were extracted from the Gimex header that was released with
the Command & Conquer Generals source code:

https://github.com/electronicarts/CnC_Generals_Zero_Hour/blob/main/Generals/Code/Libraries/Source/Compression/EAC/gimex.h

The version in that header (3.46) is one revision newer than what was
used in SimCity 4 (3.45), but the only change in 3.46 appears to have
been defining a new bit field in space that was already reserved.